+++ src/Products/urban/GenericLicence.py
@@ -366,6 +366,7 @@ schema = Schema((
         schemata='urban_road',
         default_output_type='text/plain',
     ),
+<<<<<<< .courant
     LinesField(
         name='sevesoSite',
         widget=MultiSelectionWidget(
@@ -379,6 +380,21 @@ schema = Schema((
         vocabulary=UrbanVocabulary('seveso_site', inUrbanConfig=False),
         default_method='getDefaultValue',
     ),
+=======
+    LinesField(
+        name='natura2000',
+        widget=MultiSelectionWidget(
+            format='checkbox',
+            label='natura2000',
+            label_msgid='urban_label_natura2000',
+            i18n_domain='urban',
+        ),
+        schemata='urban_road',
+        multiValued=1,
+        vocabulary=UrbanVocabulary('natura_2000', inUrbanConfig=False),
+        default_method='getDefaultValue',
+    ),
+>>>>>>> .fusion-droit.r22668
     StringField(
         name='floodingLevel',
         widget=SelectionWidget(